server/requirements.txt
adbutils==0.8.2
apkutils2==1.0.0
appdirs==1.4.4
asgiref==3.3.4
blinker==1.4
Brotli==1.0.9
cached-property==1.5.2
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
cigam==0.0.3
click==7.1.2
crefi==2.0.9
cryptography==3.3.2
decorator==4.4.2
Deprecated==1.2.10
deprecation==2.1.0
distlib==0.3.1
facebook-wda==1.4.3
filelock==3.0.12
Flask==1.1.2
Flask-JSON==0.3.4
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kaitaistruct==0.9
ldap3==2.8.1
logzero==1.6.2
lxml==4.6.1
MarkupSafe==1.1.1
mitmproxy==6.0.2
msgpack==1.0.2
numpy==1.21.0
packaging==20.4
passlib==1.7.4
Pillow==8.0.1
progress==1.5
protobuf==3.14.0
publicsuffix2==2.20191221
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pyelftools==0.27
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pytesseract==0.3.6
pyxattr==0.7.2
requests==2.25.0
retry==0.9.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
sortedcontainers==2.3.0
tornado==6.1
uiautomator2==2.12.0
urllib3==1.26.1
urwid==2.1.2
virtualenv==20.3.0
websockets==8.1
weditor==0.6.4
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.12.1
wsproto==1.0.0
xmltodict==0.12.0
zstandard==0.14.1
